\name{fpdemogpc}
\alias{fpdemogpc}
\docType{data}
\title{A sample of California wildfire polygons}
\description{This is a sample of four California wildfire polygons taken from a much larger dataset of 15491 wildfires that occurred in California.  Included to allow demonstration of various package functions while keeping the size of required package data to a minimum. 
}
\usage{data(fpdemogpc)}
\format{
A list of length 4, containing four polygons of format \code{gpc.poly}.
}
\details{
 Includes four fires: Union, Piru, Cedar and Border.
}
\source{
Contact Anthony L. Westerling for the original datafile from which this list samples: 
http://ulmo.ucmerced.edu/~westerling/index.html
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(fpdemogpc)
%## maybe str(fpdemogpc) ; plot(fpdemogpc) ...
}
\keyword{datasets}
